paremini töötavad requirements loendur/happiness + natuke kujundust ka

+++ Assets/Scripts/SceneData.cs
@@ -6,5 +6,5 @@ public class SceneData : MonoBehaviour
 {
     public static CustomerData CurrentCustomerStatic;
     public static int customersTotal = 0;
-    public static int happyCustomers = 0;
+    public static double happyCustomers = 0;
 }
